add autenhication, password change, admin acount

+++ basics/ToDo/routers/todos.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 from models import Todos
 from database import SessionLocal
-from routers import auth
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +17,8 @@ def get_db():
 
 db_dependency = Annotated[Session, Depends(get_db)]
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
+
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str=Field(min_length=3, max_length=100)
@@ -25,20 +27,28 @@ class TodoRequest(BaseModel):
     
     
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, 
+                   db: db_dependency):
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 @router.get("/todo/{id}")
-async def read_todo(db: db_dependency, id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == id).first()
+async def read_todo(user: user_dependency,
+                    db: db_dependency, id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    todo_model = db.query(Todos).filter(Todos.id == id).filter(Todos.owner_id == user.get("id")).first()
     if not todo_model:
         raise HTTPException(status_code=404, detail="todo not found")
     return todo_model
 
 
 @router.post("/todo, status_code=status.HTTP_201_CREATED")
-async def create_todo(db: db_dependency, todoRequest: TodoRequest):
-    todo_model = Todos(**todoRequest.model_dump())
+async def create_todo(user: user_dependency, 
+                      db: db_dependency, 
+                      todoRequest: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    todo_model = Todos(**todoRequest.dict(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
     db.refresh(todo_model)
@@ -46,8 +56,13 @@ async def create_todo(db: db_dependency, todoRequest: TodoRequest):
 
     
 @router.put("/todo/{id}")
-async def update_todo(db: db_dependency, todoRequest: TodoRequest, id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == id).first()
+async def update_todo(user: user_dependency,
+                      db: db_dependency, 
+                      todoRequest: TodoRequest, 
+                      id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    todo_model = db.query(Todos).filter(Todos.id == id).filter(Todos.owner_id == user.get("id")).first()
     if not todo_model:
         raise HTTPException(status_code=404, detail="todo not found")
     todo_model.title = todoRequest.title
@@ -60,8 +75,13 @@ async def update_todo(db: db_dependency, todoRequest: TodoRequest, id: int=Path(
     return todo_model
 
 @router.delete("/todo/{id}")
-async def delete_todo(db: db_dependency, todoRequest: TodoRequest, id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == id).first()
+async def delete_todo(user: user_dependency, 
+                      db: db_dependency, 
+                    #   todoRequest: TodoRequest, 
+                      id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    todo_model = db.query(Todos).filter(Todos.id == id).filter(Todos.owner_id == user.get("id")).first()
     if not todo_model:
         raise HTTPException(status_code=404, detail="todo not found")
     db.query(Todos).filter(Todos.id == id).delete()
